Prefix HTTP removed from responses.
