+++ lib/include/copper/components/response_not_found.hpp
@@ -1,5 +1,5 @@
-#ifndef COPPER_COMPONENTS_HTTP_RESPONSE_BAD_REQUEST_HPP
-#define COPPER_COMPONENTS_HTTP_RESPONSE_BAD_REQUEST_HPP
+#ifndef COPPER_COMPONENTS_RESPONSE_NOT_FOUND_HPP
+#define COPPER_COMPONENTS_RESPONSE_NOT_FOUND_HPP
 
 #pragma once
 
@@ -8,22 +8,23 @@
 #include <copper/components/shared.hpp>
 
 namespace copper::components {
+
 /**
  * Forward state
  */
 class state;
 
 /**
- * Generates HTTP response on bad requests
+ * Generate response on not found
  *
  * @param request
  * @param start_at
  * @param state
  * @return response
  */
-response http_response_bad_request(const request& request,
-                                   long start_at,
-                                   const shared<state>& state);
+response response_not_found(const request& request,
+                            long start_at,
+                            const shared<state>& state);
 
 }  // namespace copper::components
