+++ lib/include/copper/components/response_too_many_requests.hpp
@@ -1,5 +1,5 @@
-#ifndef COPPER_COMPONENTS_HTTP_RESPONSE_UNAUTHORIZED_HPP
-#define COPPER_COMPONENTS_HTTP_RESPONSE_UNAUTHORIZED_HPP
+#ifndef COPPER_COMPONENTS_RESPONSE_TOO_MANY_REQUESTS_HPP
+#define COPPER_COMPONENTS_RESPONSE_TOO_MANY_REQUESTS_HPP
 
 #pragma once
 
@@ -15,15 +15,16 @@ namespace copper::components {
 class state;
 
 /**
- * Generate HTTP response on unauthorized
+ * Generate response on rate limit exceeded
  *
  * @param request
  * @param start_at
+ * @param ttl
  * @param state
  * @return response
  */
-response http_response_unauthorized(const request& request,
-                                    long start_at,
+response response_too_many_requests(const request& request,
+                                    long start_at, int ttl,
                                     const shared<state>& state);
 
 }  // namespace copper::components
